e2e: Move `Custom SriovNetworkNodePolicy` / `Configuration` to its own file
To improve readability, reduce the sizeof `test_sriov_operator.go` by moving part of the
test cases to a different file.
It's important to keep the same test name, which are composed by the combination of each
parent `Context`/`Describe`/`It` clause.
